continuous qr code + handle import qrcode
