+++ collect_app/src/main/java/org/odk/collect/android/preferences/qr/ShowQRCodeFragment.java
@@ -12,13 +12,8 @@
 
 package org.odk.collect.android.preferences.qr;
 
-import android.app.Activity;
 import android.app.AlertDialog;
 import android.content.Context;
-import android.content.Intent;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-import android.net.Uri;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -31,26 +26,16 @@ import android.widget.TextView;
 import androidx.annotation.Nullable;
 import androidx.appcompat.widget.AppCompatCheckedTextView;
 
-import com.google.zxing.ChecksumException;
-import com.google.zxing.FormatException;
-import com.google.zxing.NotFoundException;
-
 import org.odk.collect.android.R;
 import org.odk.collect.android.analytics.Analytics;
 import org.odk.collect.android.analytics.AnalyticsEvents;
 import org.odk.collect.android.injection.DaggerUtils;
 import org.odk.collect.android.preferences.AdminSharedPreferences;
 import org.odk.collect.android.preferences.GeneralSharedPreferences;
-import org.odk.collect.android.preferences.utilities.SettingsUtils;
-import org.odk.collect.android.utilities.QRCodeUtils;
-import org.odk.collect.android.utilities.ToastUtils;
 
-import java.io.IOException;
-import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.zip.DataFormatException;
 
 import javax.inject.Inject;
 
@@ -70,8 +55,6 @@ import static org.odk.collect.android.preferences.AdminKeys.KEY_ADMIN_PW;
 import static org.odk.collect.android.preferences.GeneralKeys.KEY_PASSWORD;
 
 public class ShowQRCodeFragment extends Fragment {
-
-    private static final int SELECT_PHOTO = 111;
 
     private final CompositeDisposable compositeDisposable = new CompositeDisposable();
     private final boolean[] checkedItems = {true, true};
@@ -201,44 +184,6 @@ public class ShowQRCodeFragment extends Fragment {
         }
 
         dialog.show();
-    }
-
-    @Override
-    public void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-
-        if (requestCode == SELECT_PHOTO) {
-            if (resultCode == Activity.RESULT_OK) {
-                try {
-                    boolean qrCodeFound = false;
-                    final Uri imageUri = data.getData();
-                    if (imageUri != null) {
-                        final InputStream imageStream = getActivity().getContentResolver()
-                                .openInputStream(imageUri);
-
-                        final Bitmap bitmap = BitmapFactory.decodeStream(imageStream);
-                        if (bitmap != null) {
-                            String response = QRCodeUtils.decodeFromBitmap(bitmap);
-                            if (response != null) {
-                                qrCodeFound = true;
-                                SettingsUtils.applySettings(getActivity(), response);
-                            }
-                        }
-                    }
-                    if (!qrCodeFound) {
-                        ToastUtils.showLongToast(R.string.qr_code_not_found);
-                    }
-                } catch (FormatException | NotFoundException | ChecksumException e) {
-                    Timber.i(e);
-                    ToastUtils.showLongToast(R.string.qr_code_not_found);
-                } catch (DataFormatException | IOException | OutOfMemoryError | IllegalArgumentException e) {
-                    Timber.e(e);
-                    ToastUtils.showShortToast(getString(R.string.invalid_qrcode));
-                }
-            } else {
-                Timber.i("Choosing QR code from sdcard cancelled");
-            }
-        }
     }
 
     private Collection<String> getSelectedPasswordKeys() {
